backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
pandas==2.2.0
Flask-Limiter==3.8.0
marshmallow==3.21.3
